% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianCenter.R
\name{median_center}
\alias{median_center}
\title{Median Center}
\usage{
median_center(x, group = NULL, weight = NULL)
}
\arguments{
\item{x}{Input POINT, MULTIPOINT, POLYGON, or MULTIPOLYGON
simple features}

\item{group}{specifies groups to calculate individual mean
centers for}

\item{weight}{numeric; weight specifying an individual point's
contribution to the mean center}
}
\value{
An sf object with a mean center for each group
}
\description{
\code{median_center()} returns the Euclidean median within
a specified tolerance for each group using the method
developed by Kuhn and Kuenne 1962.
}
\section{Unprojected data}{

If \code{st_is_longlat(x)}, median center is calculated
assuming a spherical Earth.
}

\section{Projected data}{

If \code{!st_is_longlat(x)}, median center is calculated assuming
a "flat" Earth.
}

\examples{
df <- data.frame(
  lon = c(-88, -90, -92, -89, -90),
  lat = c(42, 40, 30, 32, 42),
  grp = c("a", "b", "a", "b", "a"),
  wt = c(1,1,1,1,1)) 
x_sf <- sf::st_as_sf(x, coords = c("lon", "lat"), crs = 4326)
x_transformed <- sf::st_transform(x, crs = "ESRI:102003")
median_center(x_transformed, group = "grp", weight = "wt")
}
